chore(subgraphs/dex): refactor shared entities
